refactored lerp
