+++ doc/creating_planner_plugin/include/lerp_interface/lerp_planning_context.h
@@ -1,9 +1,9 @@
 #ifndef LERP_PLANNING_CONTEXT_H
 #define LERP_PLANNING_CONTEXT_H
 
-#include <moveit/planning_interface/planning_request.h>
-#include <moveit/planning_interface/planning_response.h>
 #include <moveit/planning_interface/planning_interface.h>
+
+#include "lerp_interface/lerp_interface.h"
 
 namespace lerp_interface
 {
@@ -24,18 +24,10 @@ public:
   bool terminate() override;
   void clear() override;
 
-
-  static void plotVector(const std::string &str, const std::vector<double>& v);
-
 private:
-  trajectory_msgs::JointTrajectory interpolateMultDOF(const std::vector<double>& v1, const std::vector<double>& v2, const int& num);
-  std::vector<double> interpolateSingleDOF(const double& d1, const double& d2, const int& num);
-  int dof;
-
-
   moveit::core::RobotModelConstPtr robot_model_;
   robot_state::RobotStatePtr robot_state_;
-
+  LERPInterfacePtr lerp_interface_;
 };
 
 } // namespace lerp_interface
