+++ doc/creating_planner_plugin/src/lerp_interface.cpp
@@ -1,0 +1,129 @@
+/*********************************************************************
+ * Software License Agreement (BSD License)
+ *
+ *  Copyright (c) 2019, PickNik, Inc.
+ *  All rights reserved.
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions
+ *  are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *   * Redistributions in binary form must reproduce the above
+ *     copyright notice, this list of conditions and the following
+ *     disclaimer in the documentation and/or other materials provided
+ *     with the distribution.
+ *   * Neither the name of PickNik nor the names of its
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ *  FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ *  COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ *  BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ *  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ *  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ *  LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ *  POSSIBILITY OF SUCH DAMAGE.
+ *********************************************************************/
+
+/* Author: Omid Heidari */
+
+#include <moveit/planning_interface/planning_interface.h>
+#include <moveit/planning_scene/planning_scene.h>
+#include <moveit/robot_state/conversions.h>
+
+#include <moveit_msgs/MotionPlanRequest.h>
+
+#include <ros/ros.h>
+
+#include <limits>
+#include <vector>
+#include <Eigen/Geometry>
+#include <unordered_map>
+
+#include "lerp_interface/lerp_interface.h"
+
+namespace lerp_interface
+{
+LERPInterface::LERPInterface(const ros::NodeHandle& nh) : nh_(nh), name_("LERPInterface")
+{
+}
+
+bool LERPInterface::solve(const planning_scene::PlanningSceneConstPtr& planning_scene,
+                             const planning_interface::MotionPlanRequest& req,
+                             moveit_msgs::MotionPlanDetailedResponse& res)
+{
+  // Load the planner-specific parameters
+  nh_.getParam("num_steps", num_steps_);
+
+  ros::WallTime start_time = ros::WallTime::now();
+  robot_model::RobotModelConstPtr robot_model = planning_scene->getRobotModel();
+  robot_state::RobotStatePtr start_state(new robot_state::RobotState(robot_model));
+  *start_state = planning_scene->getCurrentState();
+  const robot_state::JointModelGroup* joint_model_group = start_state->getJointModelGroup(req.group_name);
+  std::vector<std::string> joint_names = joint_model_group->getVariableNames();
+  int dof = joint_names.size();
+  std::vector<double> start_joint_values;
+  start_state->copyJointGroupPositions(joint_model_group, start_joint_values); ; //req.start_state.joint_state.position;
+
+  // This planner only supports one goal constriant in the request
+  std::vector<moveit_msgs::Constraints> goal_constraints = req.goal_constraints;
+  std::vector<moveit_msgs::JointConstraint> goal_joint_constraint = goal_constraints[0].joint_constraints;
+
+  std::vector<double> goal_joint_values;
+  for (auto x : goal_joint_constraint)
+  {
+    goal_joint_values.push_back(x.position);
+  }
+
+  // ==================== Interpolation
+  trajectory_msgs::JointTrajectory joint_trajectory;
+  joint_trajectory.points.resize(num_steps_ + 1);
+
+  std::vector<double> dt_vector;
+  for (int joint_index = 0; joint_index < dof; ++joint_index){
+    double dt = ( goal_joint_values[joint_index] - start_joint_values[joint_index] ) / num_steps_;
+    dt_vector.push_back(dt);
+  }
+
+  robot_state::RobotStatePtr robot_state(start_state);
+  for (int step = 0; step <= num_steps_ ; ++step)
+    {
+      std::vector<double> joint_values;
+      for (int k = 0; k < dof; ++k){
+        double joint_value = start_joint_values[k] + step * dt_vector[k];
+        joint_values.push_back(joint_value);
+      }
+      robot_state->setJointGroupPositions(joint_model_group, joint_values);
+      robot_state->update();
+
+      if(!planning_scene->isStateValid(*robot_state, req.group_name, false))
+        ROS_FATAL_STREAM_NAMED(name_, "robot is in collision at step: " << step);
+
+      joint_trajectory.joint_names = joint_names;
+      joint_trajectory.points[step].positions = joint_values;
+    }
+
+  // ==================== feed the response
+  res.trajectory.resize(1);
+  res.trajectory[0].joint_trajectory.joint_names = joint_names;
+  res.trajectory[0].joint_trajectory.header = req.start_state.joint_state.header;
+  res.trajectory[0].joint_trajectory = joint_trajectory;
+
+  res.error_code.val = moveit_msgs::MoveItErrorCodes::SUCCESS;
+  res.processing_time.push_back((ros::WallTime::now() - start_time).toSec());
+
+  res.group_name = req.group_name;
+  res.trajectory_start.joint_state.name = joint_names;
+  res.trajectory_start.joint_state.position = start_joint_values;
+
+  return true;
+}
+
+}  // namespace trajopt_interface
